+++ src/main/java/com/sa/baff/model/dto/GoalsDto.java
@@ -22,9 +22,9 @@ public class GoalsDto {
         /* 목표 기간(종료일) */
         private LocalDateTime endDate;
         /* 시작 몸무게 */
-        private Integer startWeight;
+        private Double startWeight;
         /* 목표 몸무게 */
-        private Integer targetWeight;
+        private Double targetWeight;
         /* 목표기간 만료여부 */
         private Boolean isExpired;
